Restore round time on app load
If the time between the last run and now is less than the time left in
the round when we last closed the app, subtract (now - last run) from
time left in round.

+++ src/layout.h
@@ -21,4 +21,4 @@ void layout_group_mark_dirty(LayoutGroup *layout_group);
 void layout_group_add_to_window(LayoutGroup *layout_group, Window *window);
 void layout_group_select_player(LayoutGroup *layout_group, Player *player);
 void layout_group_update_player(LayoutGroup *layout_group, Player *player);
-long layout_group_round_time_tick(LayoutGroup *layout_group);
+uint32_t layout_group_round_time_tick(LayoutGroup *layout_group);
