+++ src/round_time_layer.h
@@ -7,12 +7,12 @@
 typedef struct RoundTimeLayer {
     Layer *layer;
     TextLayer *text_layer;
-    long time_left;
+    uint32_t time_left;
 } RoundTimeLayer;
 
 RoundTimeLayer *round_time_layer_create();
 void round_time_layer_destroy(RoundTimeLayer *round_time_layer);
 Layer *round_time_layer_get_layer(RoundTimeLayer *round_time_layer);
 void round_time_layer_set_frame(RoundTimeLayer *round_time_layer, GRect frame);
-void round_time_layer_set_time_left(RoundTimeLayer *round_time_layer, long time_left);
-long round_time_layer_tick(RoundTimeLayer *round_time_layer);
+void round_time_layer_set_time_left(RoundTimeLayer *round_time_layer, uint32_t time_left);
+uint32_t round_time_layer_tick(RoundTimeLayer *round_time_layer);
